requirements.txt
gym~=0.17.3
numpy==1.19.4
boolean.py~=3.7
networkx==2.4
pyyaml~=5.3.1
setuptools~=49.2.1
matplotlib~=3.2.1
plotly~=4.11.0
tabulate~=0.8.7
ordered_set==4.0.2
progressbar2==3.51.4
